% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_after.R
\name{cleanup_after}
\alias{cleanup_after}
\title{Cleanup after unzipping kmz}
\usage{
cleanup_after(files)
}
\arguments{
\item{files}{a character vector of files to delete}
}
\value{
nothing
}
\description{
This function is used internally and shouldn't normally be called on it's own.
}
