% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_kml.R
\name{handle_kml}
\alias{handle_kml}
\title{Load .kml files into R}
\usage{
handle_kml(kml)
}
\arguments{
\item{kml}{a file path to a valid .kml file}
}
\value{
an sf object of the kml file from the input kml
}
\description{
Use sf to read .kml files. Handles multi-layer files. This function is
used internally and shouldn't normally be called
}
