% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_kmz.R
\name{handle_kmz}
\alias{handle_kmz}
\title{Read .kmz file into R}
\usage{
handle_kmz(kmz)
}
\arguments{
\item{kmz}{a file path to a valid .kmz file}
}
\value{
an sf object of the input .kmz file
}
\description{
Use sf to read .kmz files. Handles multi-layer files. This function is
used internally and shouldn't normally be called
}
