% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_keyhole.R
\name{read_keyhole}
\alias{read_keyhole}
\title{Read simple features from kml or kmz files}
\usage{
read_keyhole(file)
}
\arguments{
\item{file}{a file path to a valid .kml or .kmz file}
}
\value{
an sf object when the layer is successfully read.
}
\description{
Read simple features from kml or kmz files
}
